third_party/WebKit/LayoutTests/geolocation-api/timeout-clear-watch.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/geolocation-mock.js"></script>
</head>
<body>
<script>
description("Tests that when a watch times out and is cleared from the error callback, there is no crash. This a regression test for https://bugs.webkit.org/show_bug.cgi?id=32111.");

var error;

geolocationMock.then(mock => {
    mock.setGeolocationPosition(51.478, -0.166, 100.0);

    var watchId = navigator.geolocation.watchPosition(function() {
        testFailed('Success callback invoked unexpectedly');
        finishJSTest();
    }, function(e) {
        error = e;
        shouldBe('error.code', 'error.TIMEOUT');
        shouldBe('error.message', '"Timeout expired"');
        navigator.geolocation.clearWatch(watchId);
        window.setTimeout(finishJSTest, 0);
    }, {
        timeout: 0
    });
});

window.jsTestIsAsync = true;
</script>
</body>
</html>
